+++ rest_framework/__init__.py
@@ -1,3 +1,3 @@
-__version__ = '2.1.9'
+__version__ = '2.1.10'
 
 VERSION = __version__  # synonym